# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "jaguar_food_app",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "variables": {}
    }
  ],
  "paths": {
    "/auth/user/signup": {
      "post": {
        "tags": ["authfolder"],
        "summary": "createAdmin",
        "description": "this creates an admin user",
        "operationId": "createAdmin",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/createAdminRequest"
                  },
                  {
                    "example": {
                      "email": "test99@gmail.com",
                      "password": "test999",
                      "first_name": "test4",
                      "last_name": "test4",
                      "phone_number": "12345678912",
                      "organization_name": "test99",
                      "lunch_price": 300,
                      "currency": "tetst4",
                      "currency_code": "NGN"
                    }
                  }
                ]
              },
              "example": {
                "email": "test99@gmail.com",
                "password": "test999",
                "first_name": "test4",
                "last_name": "test4",
                "phone_number": "12345678912",
                "organization_name": "test99",
                "lunch_price": 300,
                "currency": "tetst4",
                "currency_code": "NGN"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": []
      }
    },
...
    "/auth/organization/{name}": {
      "get": {
        "tags": ["authfolder"],
        "summary": "searchOrg",
        "description": "searches the name of an organization",
        "operationId": "searchOrg",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "test99"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/orgs/invite": {
      "post": {
        "tags": ["adminActivitiesFolder"],
        "summary": "createInvite",
        "description": "allows an admin to send an invitation link to specified email",
        "operationId": "createInvite",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/createInviteRequest"
                  },
                  {
                    "example": {
                      "email": "mhyke.badmus@gmail.com"
                    }
                  }
                ]
              },
              "example": {
                "email": "mhyke.badmus@gmail.com"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/orgs/lunch/update": {
      "patch": {
        "tags": ["adminActivitiesFolder"],
        "summary": "update0rgFoodPrice",
        "description": "allows an admin to update his/her organization's food price",
        "operationId": "update0rgFoodPrice",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/update0rgFoodPriceRequest"
                  },
                  {
                    "example": {
                      "lunch_price": 300
                    }
                  }
                ]
              },
              "example": {
                "lunch_price": 300
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
      "createAdminRequest": {
        "title": "createAdminRequest",
        "required": [
          "email",
          "password",
          "first_name",
          "last_name",
          "phone_number",
          "organization_name",
          "lunch_price",
          "currency",
          "currency_code"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "phone_number": {
            "type": "string"
          },
          "organization_name": {
            "type": "string"
          },
          "lunch_price": {
            "type": "integer",
            "format": "int32"
          },
          "currency": {
            "type": "string"
          },
          "currency_code": {
            "type": "string"
          }
        },
        "example": {
          "email": "test99@gmail.com",
          "password": "test999",
          "first_name": "test4",
          "last_name": "test4",
          "phone_number": "12345678912",
          "organization_name": "test99",
          "lunch_price": 300,
          "currency": "tetst4",
          "currency_code": "NGN"
        }
      },
...
      "createInviteRequest": {
        "title": "createInviteRequest",
        "required": ["email"],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "example": {
          "email": "mhyke.badmus@gmail.com"
        }
      },
...
      "update0rgFoodPriceRequest": {
        "title": "update0rgFoodPriceRequest",
        "required": ["lunch_price"],
        "type": "object",
        "properties": {
          "lunch_price": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "lunch_price": 300
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}